requirements.txt
python-daemon
jinja2
